DoddsWattsmodel

Documentation for DoddsWattsmodel.

Dodds_Watts_model.Agent

Dodds_Watts_model.fixed_point_curve_cmm

Dodds_Watts_model.fixed_point_curve_etm

Dodds_Watts_model.initialize_cmm

Dodds_Watts_model.initialize_etm

Dodds_Watts_model.initialize_universal

Dodds_Watts_model.plot_model_cmm

Dodds_Watts_model.plot_model_etm

Dodds_Watts_model.plot_model_network

Dodds_Watts_model.plot_model_periodic

Dodds_Watts_model.plot_model_universal

Dodds_Watts_model.plot_model_vacc

Dodds_Watts_model.plot_network

Dodds_Watts_model.plot_ode

Dodds_Watts_model.timestep_cmm

Dodds_Watts_model.timestep_etm

Dodds_Watts_model.timestep_network

Dodds_Watts_model.timestep_universal

Dodds_Watts_model.Agent — Type

Definition of the individual agents

type 1: Susceptible type 2: Infected type 3: Removed

dose: Count of received dosages

threshold: individual dose threshold; if the agents exceeds its threshold, it becomes infected

vaccine: Vaccinated yes or no

Dodds_Watts_model.fixed_point_curve_cmm — Method

Plots the fixed point curve for the critical mass model according to the memory T and threshold d

Dodds_Watts_model.fixed_point_curve_etm — Method

Plots the stable fixed point curve of the epidemic threshold model according to the memory T and recovery rate r

Dodds_Watts_model.initialize_cmm — Method

Initializes N agents with threshold d, where I are infected for the critical mass model

Dodds_Watts_model.initialize_etm — Method

Initializes N agents, where I are infected for the epidemic threshold model

Dodds_Watts_model.initialize_universal — Method

Initializes N agents, where I are infected for the epidemic threhsold model. The threshold distribution is set to Gamma(1,1)

Dodds_Watts_model.plot_model_cmm — Method

Plots a simulation of the critical mass model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, exposure probability p, memory T, recovery rate r, probabiltiy imunity loss ρ, threshold d

Dodds_Watts_model.plot_model_etm — Method

Plots a simulation of the epidemic threshold model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, exposure probability p, memory T, recovery rate r, probabiltiy imunity loss ρ

Dodds_Watts_model.plot_model_network — Method

Plots a simulation of the network model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, exposure probability p, memory T, recovery rate r, probabiltiy imunity loss ρ and network g The variable version determines the underlying model: "etm" = epidemic threshold model "ciritical" = critical mass model "universal" = heterogeneity model

Dodds_Watts_model.plot_model_periodic — Method

Plots a simulation of the periodic model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, periodic exposure distribution p(t) with period P, memory T, recovery rate r, probabiltiy imunity loss ρ The variable version determines the underlying model: "etm" = epidemic threshold model "ciritical" = critical mass model "universal" = heterogeneity model

Dodds_Watts_model.plot_model_universal — Method

Plots a simulation of the heterogeneity model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, exposure probability p, memory T, recovery rate r, probabiltiy imunity loss ρ

Dodds_Watts_model.plot_model_vacc — Method

Plots a simulation of the vaccination model for t timesteps, N individuals, S suscpetibles, I infected, R recoverd, exposure probability p, memory T, recovery rate r, probabiltiy imunity loss ρ, vaccination rate v and strength s The vaccination rate determines how many of the agents are being vaccinated. Their threshold will be raised by the vaccine strength s. The variable version determines the underlying model: "etm" = epidemic threshold model "ciritical" = critical mass model "universal" = heterogeneity model

Dodds_Watts_model.plot_network — Method

Plots an embedding of the network g

Dodds_Watts_model.plot_ode — Method

Plots the corresponding ODE to the epidemic threshold model with N individuals, exposure probability p and recovery rate r

Dodds_Watts_model.timestep_cmm — Method

Performs a time step for the critical mass model

Dodds_Watts_model.timestep_etm — Method

Performs a time step for the epidemic threshold model

Dodds_Watts_model.timestep_network — Method

Performs a time step for the Dodds-Watts model with a network g

Dodds_Watts_model.timestep_universal — Method

Performs a time step for the heterogeneity model. The dose distribution is set to LogNormal()