DoddsWattsmodel

Documentation for DoddsWattsmodel.

Dodds_Watts_model.Agent

Dodds_Watts_model.fixed_point_curve_cmm

Dodds_Watts_model.fixed_point_curve_etm

Dodds_Watts_model.initialize_cmm

Dodds_Watts_model.initialize_etm

Dodds_Watts_model.initialize_universal

Dodds_Watts_model.plot_model_cmm

Dodds_Watts_model.plot_model_etm

Dodds_Watts_model.plot_model_network

Dodds_Watts_model.plot_model_periodic

Dodds_Watts_model.plot_model_universal

Dodds_Watts_model.plot_model_vacc

Dodds_Watts_model.plot_network

Dodds_Watts_model.plot_ode

Dodds_Watts_model.timestep_cmm

Dodds_Watts_model.timestep_etm

Dodds_Watts_model.timestep_network

Dodds_Watts_model.timestep_universal

Dodds_Watts_model.Agent — Type

Definition of the individual agents

Dodds_Watts_model.fixed_point_curve_cmm — Method

Plots the fixed point curve for the critical mass model

Dodds_Watts_model.fixed_point_curve_etm — Method

plots the stable fixed point curve of the Epidemic threshold model

Dodds_Watts_model.initialize_cmm — Method

Initializes agents for the critical mass model

Dodds_Watts_model.initialize_etm — Method

Initializes agents for the epidemic threhsold model

Dodds_Watts_model.initialize_universal — Method

Initializes agents for the heterogeneity model

Dodds_Watts_model.plot_model_cmm — Method

Plots a simulation of the critical mass model for t timesteps

Dodds_Watts_model.plot_model_etm — Method

Plots a simulation of the epidemic threshold model

Dodds_Watts_model.plot_model_network — Method

Plots a simulation of the model spcified in version for t timesteps with an underlying network

Dodds_Watts_model.plot_model_periodic — Method

Plots a simulation of the model spcified in version for t timesteps with an periodic distribution function p(t)

Dodds_Watts_model.plot_model_universal — Method

Plots a simulation of the heterogeneity model for t timesteps

Dodds_Watts_model.plot_model_vacc — Method

Plots a simulation of the model spcified in version for t timesteps with a vaccination rate

Dodds_Watts_model.plot_network — Method

Plots an embedding of the graph g

Dodds_Watts_model.plot_ode — Method

plots the corresponding ODE to the epidemic threshold model

Dodds_Watts_model.timestep_cmm — Method

Performs a time step for the critical mass model

Dodds_Watts_model.timestep_etm — Method

Performs a time step for the epidemic threshold model

Dodds_Watts_model.timestep_network — Method

Performs a time step for the Dodds-Watts model with a network g

Dodds_Watts_model.timestep_universal — Method

Performs a time step for the heterogeneity model